circuit, schedule separated; tests pass
